feat: Moved Extension/Quality files to root
